requirements.txt
fastapi
uvicorn
langchain
langgraph
google-generativeai
sqlmodel
psycopg2-binary
httpx
python-dotenv
faker
gradio
passlib[bcrypt]
bcrypt==4.0.1
PyJWT
pydantic[email]
python-jose[cryptography]